refactor: move the notification inbox to the platform (#2121)
A notification inbox is per-user, not per-edition. Left in n23.core, a player
with both an n23 and an n26 gang would get two inboxes, two navbar badges with
no defined way to combine them, and every site-wide broadcast written twice.

#2118 removed the blocker by giving Notification generic `target` / `scope`
relations, but deliberately kept the `related_list` / `related_campaign` FK
columns and kept writing them so instances on the previous revision survived
the rollout. That revision is now live, so this commit finishes the job:

- Drops the two FK columns and the dual-write in `notify()`. `notify()` now
  takes `target=` / `scope=` directly. `notify_campaign_arbitrator` grows the
  same `target=` kwarg in place of `related_list=`.
- Moves the model to `gyrinx.site`, beside Banner and ImpersonationLog. State
  only: `db_table` stays `core_notification`, every index stays where it is,
  and the ContentType row is updated in place so its id — and the eight
  permissions hanging off it — survive. Unlike the earlier moves there is no
  parallel Historical* table to pin; this model declares no HistoricalRecords.
- Moves what follows the model: the `notifications` context processor (which
  clears the last `n23.` entry in the settings context-processor list, and
  empties `n23/core/context_processors.py`), the admin registration and its
  two templates, and the model-level tests.

The inbox views and their URLs stay in the edition — no `core:` URL name and no
path changes. The admin URL names do change, since Django derives them from the
app label: `admin:core_notification_*` becomes `admin:gyrinxsite_notification_*`.

Notification joins the parametrised guard in test_admin_registration: the star
imports in n23/core/admin/__init__.py are exactly how the Event admin got
silently unregistered during #2109.

Verified against a populated database, not just the suite (which runs
--nomigrations and never executes these): 46 notification rows before and after,
including 30 with the legacy FK columns set; only those two columns and their
two indexes dropped; all four named indexes intact; ContentType 224 keeps its id
with app_label core -> gyrinxsite; no gyrinxsite_* table created. Reversing and
re-applying both migrations round-trips cleanly.

Closes #2121. Refs #2093.
