+++ gyrinx/site/__init__.py
@@ -1,9 +1,12 @@
-"""Site-level platform models: the announcement banner and the impersonation log.
+"""Site-level platform models: the announcement banner, the impersonation log,
+and the per-user notification inbox.
 
-Both are about running Gyrinx, not about playing a game edition.
+All three are about running Gyrinx, not about playing a game edition. The
+inbox in particular is per-user, not per-edition: one badge, one list, one
+broadcast, whatever editions the user has gangs in.
 
-Notification deliberately is NOT here. It carries ForeignKeys to `core.List`
-and `core.Campaign`, so moving it would give a platform model hard references
-to edition tables. Generalising those two columns is design work, tracked
-separately.
+`Notification` refers to edition objects through generic `target` / `scope`
+relations rather than ForeignKeys, so nothing here has a schema dependency on
+an edition table. Its inbox *views* stay in the edition, under the `core:` URL
+names they have always had.
 """
