+++ gyrinx/site/templates/admin/gyrinxsite/notification/broadcast.html
@@ -3,7 +3,7 @@
     <div class="breadcrumbs">
         <a href="{% url 'admin:index' %}">Home</a>
         › <a href="{% url 'admin:app_list' app_label=opts.app_label %}">{{ opts.app_config.verbose_name }}</a>
-        › <a href="{% url 'admin:core_notification_changelist' %}">{{ opts.verbose_name_plural|capfirst }}</a>
+        › <a href="{% url 'admin:gyrinxsite_notification_changelist' %}">{{ opts.verbose_name_plural|capfirst }}</a>
         › Broadcast
     </div>
 {% endblock breadcrumbs %}
