+++ gyrinx/site/templates/admin/gyrinxsite/notification/change_list.html
@@ -1,7 +1,8 @@
 {% extends "admin/change_list.html" %}
 {% block object-tools-items %}
     <li>
-        <a href="{% url 'admin:core_notification_broadcast' %}" class="addlink">Broadcast</a>
+        <a href="{% url 'admin:gyrinxsite_notification_broadcast' %}"
+           class="addlink">Broadcast</a>
     </li>
     {{ block.super }}
 {% endblock object-tools-items %}
