+++ n23/core/admin/__init__.py
@@ -4,7 +4,6 @@ from .campaign import *  # noqa: F403
 from .events import *  # noqa: F403
 from .impersonation import *  # noqa: F403
 from .list import *  # noqa: F403
-from .notification import *  # noqa: F403
 from .pack import *  # noqa: F403
 from .site import *  # noqa: F403
 from .upload import *  # noqa: F403
